fix rapide

+++ routes/categories.js
@@ -13,21 +13,21 @@ router.get('/', checkToken, async (req, res) => {
     })
 })
 
-//Get 1 category by id
-router.get('/:id', async (req, res) => {
-    const sql = "SELECT * FROM categories WHERE id= ?"
-    dbQuery(sql, [req.params.id]).then((results) => {
+//Get all main category (no parents)
+router.get('/main', async (req, res) => {
+    const sql = "SELECT * FROM categories WHERE parent_id IS NULL"
+    dbQuery(sql).then((results) => {
         res.status(200).json(results)
     }).catch((error) => {
         res.status(500).send({'error': error.message})
     })
 })
 
-//Get all main category (no parents)
-router.get('/main', async (req, res) => {
-    const sql = "SELECT * FROM categories WHERE parent_id IS NULL"
-    dbQuery(sql).then((results) => {
-        res.status(200).json(results)
+//Get 1 category by id
+router.get('/:id', async (req, res) => {
+    const sql = "SELECT * FROM categories WHERE id= ?"
+    dbQuery(sql, [req.params.id]).then((results) => {
+        res.status(200).json(results[0])
     }).catch((error) => {
         res.status(500).send({'error': error.message})
     })
@@ -44,32 +44,47 @@ router.get('/from_parent/:id', async (req, res) => {
 })
 //removes 1 category (by id)
 router.delete('/:id', async (req, res) => {
-    const sql = "DELETE FROM categories WHERE id = ?"
-    dbQuery(sql, [req.params.id]).then(() => {
-        res.sendStatus(200)
-    }).catch((error) => {
-        res.status(500).send({'error': error.message})
-    })
+    const decoded = checkToken(req.body.token)
+    if (decoded.role !== 'admin') {
+        res.status(403).send({'error': 'User is not admin'})
+    } else {
+        const sql = "DELETE FROM categories WHERE id = ?"
+        dbQuery(sql, [req.params.id]).then(() => {
+            res.sendStatus(200)
+        }).catch((error) => {
+            res.status(500).send({'error': error.message})
+        })
+    }
 })
 
 //Create category
 router.post('/', async (req, res) => {
-    const sql = "INSERT INTO categories (parent_id, name) VALUES (?, ?)"
-    dbQuery(sql, [req.body.category.parent_id, req.body.category.name]).then(() => {
-        res.sendStatus(201)
-    }).catch((error) => {
-        res.status(500).send({'error': error.message})
-    })
+    const decoded = checkToken(req.body.token)
+    if (decoded.role !== 'admin') {
+        res.status(403).send({'error': 'User is not admin'})
+    } else {
+        const sql = "INSERT INTO categories (parent_id, name) VALUES (?, ?)"
+        dbQuery(sql, [req.body.category.parent_id, req.body.category.name]).then(() => {
+            res.sendStatus(201)
+        }).catch((error) => {
+            res.status(500).send({'error': error.message})
+        })
+    }
 })
 
 //Updates a category
 router.put('/:id', async (req, res) => {
-    const sql = "UPDATE categories SET name = ?, parent_id = ? WHERE id = ? "
-    dbQuery(sql, [req.body.category.name, req.body.category.parent_id, req.params.id]).then(() => {
-        res.sendStatus(200)
-    }).catch((error) => {
-        res.status(500).send({'error': error.message})
-    })
+    const decoded = checkToken(req.body.token)
+    if (decoded.role !== 'admin') {
+        res.status(403).send({'error': 'User is not admin'})
+    } else {
+        const sql = "UPDATE categories SET name = ?, parent_id = ? WHERE id = ? "
+        dbQuery(sql, [req.body.category.name, req.body.category.parent_id, req.params.id]).then(() => {
+            res.sendStatus(200)
+        }).catch((error) => {
+            res.status(500).send({'error': error.message})
+        })
+    }
 })
 
 module.exports = router
