+++ routes/products.js
@@ -127,21 +127,19 @@ router.get('/:id', async (req, res) => {
  *         description: Erreur serveur
  */
 router.post('/', async (req, res) => {
-    /*
-    try {
-        checkToken(req.body.token)
-    } catch (error) {
-        throw error
+    const decoded = checkToken(req.body.token)
+    if (decoded.role !== 'admin') {
+        res.sendStatus(403)
+    } else {
+        const product = req.body.product
+        const values = [product.name, product.description, product.price, product.images]
+        const sql = "INSERT INTO products (name, description, price, images) VALUES (?,?,?,?)"
+        dbQuery(sql, values).then(() => {
+            res.sendStatus(201)
+        }).catch((error) => {
+            res.status(500).send({'error': error.message})
+        });
     }
-     */
-    const product = req.body.product
-    const values = [product.name, product.description, product.price, product.images]
-    const sql = "INSERT INTO products (name, description, price, images) VALUES (?,?,?,?)"
-    dbQuery(sql, values).then(() => {
-        res.sendStatus(201)
-    }).catch((error) => {
-        res.status(500).send({'error': error.message})
-    });
 });
 
 /**
@@ -166,19 +164,17 @@ router.post('/', async (req, res) => {
  *         description: Erreur serveur
  */
 router.delete('/:id', async (req, res) => {
-    /*
-    try {
-        checkToken(req.body.token)
-    } catch (error) {
-        throw error
+    const decoded = checkToken(req.body.token)
+    if (decoded.role !== 'admin') {
+        res.sendStatus(403)
+    } else {
+        const sql = 'DELETE FROM products WHERE id = ?'
+        dbQuery(sql, [req.params.id]).then((results) => {
+            results.affectedRows === 0 ? res.sendStatus(404) : res.sendStatus(200)
+        }).catch((error) => {
+            res.status(500).send({'error': error.message})
+        })
     }
-    */
-    const sql = 'DELETE FROM products WHERE id = ?'
-    dbQuery(sql, [req.params.id]).then((results) => {
-        results.affectedRows === 0 ? res.sendStatus(404) : res.sendStatus(200)
-    }).catch((error) => {
-        res.status(500).send({'error': error.message})
-    })
 })
 
 /**
@@ -225,22 +221,26 @@ router.delete('/:id', async (req, res) => {
  *         description: Erreur serveur
  */
 router.put('/:id', async (req, res) => {
-    const product = req.body.product;
-    const values = [product.name, product.description, product.price, product.images, req.params.id]
-    const sql = 'UPDATE products SET name = ?, description = ?, price = ?, images = ? WHERE id = ?'
-    dbQuery(sql, values).then((results) => {
-        results.length === 0 ? res.sendStatus(404) : res.sendStatus(200)
-    }).catch((error) => {
-        res.status(500).send({'error': error.message})
-    })
+    const decoded = checkToken(req.body.token)
+    if (decoded.role !== 'admin') {
+        res.sendStatus(403)
+    } else {
+        const product = req.body.product;
+        const values = [product.name, product.description, product.price, product.images, req.params.id]
+        const sql = 'UPDATE products SET name = ?, description = ?, price = ?, images = ? WHERE id = ?'
+        dbQuery(sql, values).then((results) => {
+            results.length === 0 ? res.sendStatus(404) : res.sendStatus(200)
+        }).catch((error) => {
+            res.status(500).send({'error': error.message})
+        })
+    }
 })
 
 //Routes for fetching products by their categories
 router.get('/category/:id', async (req, res) => {
     const sql = "SELECT product_id FROM products_categories WHERE category_id = ?"
     const response = []
-    dbQuery(sql, [req.params.id]).then(async (results) => {
-        //TODO: make request for each product
+    await dbQuery(sql, [req.params.id]).then(async (results) => {
         for (const result of results) {
             const sqlEachProduct = "SELECT * FROM products WHERE id = ?"
             await dbQuery(sqlEachProduct, [result.product_id]).then(async (resultsEachProduct) => {
@@ -256,21 +256,31 @@ router.get('/category/:id', async (req, res) => {
 })
 //Sets a category for a product (both id)
 router.post('/set_category/:id', async (req, res) => {
-    const sql = "INSERT INTO products_categories (product_id, category_id) VALUES (?,?)"
-    dbQuery(sql, [req.params.id, req.body.category_id]).then(async () => {
-        res.sendStatus(201)
-    }).catch((error) => {
-        res.status(500).send({'error': error.message})
-    })
+    const decoded = checkToken(req.body.token)
+    if (decoded.role !== 'admin') {
+        res.sendStatus(403)
+    } else {
+        const sql = "INSERT INTO products_categories (product_id, category_id) VALUES (?,?)"
+        dbQuery(sql, [req.params.id, req.body.category_id]).then(async () => {
+            res.sendStatus(201)
+        }).catch((error) => {
+            res.status(500).send({'error': error.message})
+        })
+    }
 })
 
 //Removes a category from a product (both id)
 router.delete('/remove-category/:id', async (req, res) => {
-    const sql = 'DELETE FROM products_categories WHERE product_id = ? AND category_id = ?'
-    dbQuery(sql, [req.params.id, req.body.category_id]).then(() => {
-        res.sendStatus(200)
-    }).catch((error) => {
-        res.status(500).send({'error': error.message})
-    })
+    const decoded = checkToken(req.body.token)
+    if (decoded.role !== 'admin') {
+        res.sendStatus(403)
+    } else {
+        const sql = 'DELETE FROM products_categories WHERE product_id = ? AND category_id = ?'
+        dbQuery(sql, [req.params.id, req.body.category_id]).then(() => {
+            res.sendStatus(200)
+        }).catch((error) => {
+            res.status(500).send({'error': error.message})
+        })
+    }
 })
 module.exports = router
